bsp: simplify nrf53 specific implementation of tdma client/server library
